services/ml-facial/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pillow==10.1.0
python-multipart==0.0.6